Merge tf to maciekcc

+++ tensorflow/core/common_runtime/process_util.h
@@ -35,7 +35,7 @@ void SchedClosure(std::function<void()> closure);
 
 // Schedule "closure" after the given number of microseconds in the
 // fixed-size ThreadPool used for non-blocking compute tasks.
-void SchedNonBlockingClosureAfter(int micros, std::function<void()> closure);
+void SchedNonBlockingClosureAfter(int64 micros, std::function<void()> closure);
 
 }  // namespace tensorflow
