+++ tensorflow/core/kernels/sparse_dense_binary_op_shared.cc
@@ -91,7 +91,7 @@ class SparseDenseBinaryOpShared : public OpKernel {
     auto VecGreaterEq = [](ArraySlice<int64> lhs, ArraySlice<int64> rhs) {
       if (lhs.size() > rhs.size()) return true;
       if (lhs.size() < rhs.size()) return false;
-      for (int i = 0; i < lhs.size(); ++i) {
+      for (size_t i = 0; i < lhs.size(); ++i) {
         if (lhs[i] < rhs[i]) return false;
       }
       return true;
@@ -163,6 +163,10 @@ class SparseDenseBinaryOpShared : public OpKernel {
   }
 };
 
+// NOTE(aselle): If Div is extended to non-reals, make sure to use the same
+// separation of operator semantics as done for dense cwise ops. I.e. you
+// should make SparseDenseCwiseRealDiv, SparseDenseCwiseTruncateDiv,
+// SparseDenseCwiseFloorDiv, and then deprecate, SparseDenseCwiseDiv.
 // TODO(zongheng): extend to other eligible cwise operations as requested.
 #define REGISTER_KERNELS(T)                                                  \
   REGISTER_KERNEL_BUILDER(                                                   \
